src/mAIgic_integrations/email/email_gmail_impl/pyproject.toml
[project]
name = "maigic-email-gmail"
version = "0.1.0"
description = "mAIgic Gmail Implementation - Google-specific email integration"
requires-python = ">=3.12"
dependencies = [
    "maigic-email-api>=0.1.0",
    "google-api-python-client>=2.175.0",
    "google-auth>=2.40.0",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.0",
    "python-dotenv>=1.0.0",
]

# Workspace sources are defined in root pyproject.toml

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = ["-v"]
asyncio_mode = "auto"